refactor(cloudkit): split SyncCoordinator (1365→≤334) (#272, #268, #269, #275)
Clear the ERROR-level file_length on the biggest sync-engine file
(1365 > 1000) without changing behaviour. Extract methods to six
sibling `SyncCoordinator+*.swift` extension files grouped by concern
(Lifecycle, Zones, Backfill, RecordChanges, Delegate, Refetch). The
canonical class file keeps the @MainActor class declaration, nested
types (ZoneType, ObserverToken, BatchKind, PreparedEngine), stored
members, init, observer pattern, handler access, and state-file
persistence — down to 334 lines from 1365.

Also clears the baselined body-length / complexity entries keyed to
the old file by extracting helpers in the three hot-spot methods:

  - handleFetchedRecordZoneChangesAsync: 85 → 22 body, cx 16 → ~3 via
    applyFetchedZoneChanges / applyFetchedIndexChanges /
    applyFetchedProfileDataChanges.
  - handleSentRecordZoneChanges: 64 → 34 body, cx 11 → ~5 via
    processSentZone / updateQuotaExceededState.
  - nextRecordZoneChangeBatchOnMain: 101 → 38 body, cx 26 → ~6 via
    dedupedPendingChanges / buildRecordsToSave /
    appendProfileIndexRecords / appendProfileDataRecords.

Visibility: the extensions need sibling access to a number of stored
members, widened from `private` to module-internal per `strict_fileprivate`
(`fileprivate` is disallowed). Each widening is documented in-place
with a `// internal (was private) so sibling extension files can …`
comment. Four `private(set) var` stored properties (`isRunning`,
`isFetchingChanges`, `isQuotaExceeded`, `refetchAttempts`) lose their
`private(set)` because extensions in separate files aren't inside the
`private(set)` scope — external callers only read these, so the
observable surface is unchanged.

Signpost / slow-zone timing, quota-tracking order, and `continue` vs
early-return semantics verified byte-identical to the pre-split version
during extraction.

Stacked on #356 until it lands; retarget to main then.
